Remove memoizer from dartdoc (never wound up being viable) and clean up grind so it depends on less
